requirements.txt
pandas
vaderSentiment
psycopg2-binary
textblob